Fixed data and added favicons
